Organized routes in server
