+++ src/leo-server-app/src/types/schedule.ts
@@ -1,0 +1,4 @@
+export enum ScheduleType {
+  "LIVE" = "LIVE",
+  "FUTURE" = "FUTURE",
+}
