+++ src/leo-server-app/src/types/user.ts
@@ -1,0 +1,4 @@
+export enum UserRole {
+  "OPERATOR" = "OPERATOR",
+  "ADMIN" = "ADMIN",
+}
